edited existing tests to ensure more consistent passing
increased timers for web drivers, giving the program more time before
timing out (#36)
bound a personnel object to pwtestuser1 in the cucumber password change
tests. Previously, the tests would fail because the system could not
find a patient or personnel (the classes that store email addresses) for
the user pwtestuser1. I don't know if the system wants you to send an
email every time you change password, but I'm going to assume so and
ensure that there is a personnel/patient for anyone that tries to change
their password. (#28)